Move extract_int into InvenTree.helpers
